Silence Hashie warnings
Hashie is raising the warning `WARN -- : You are setting a key that conflicts with a built-in method Hashie::Mash#keys defined in Hash. This can cause unexpected behavior when accessing the key via as a property. You can still access the key via the #[] method.` more than 30 times, flooding the logs.